+++ src/rtc/rtc_connection.cpp
@@ -4,6 +4,32 @@
 
 namespace sora {
 
+// stats のコールバックを受け取るためのクラス
+class RTCStatsCallback : public webrtc::RTCStatsCollectorCallback {
+ public:
+  typedef std::function<void(
+      const rtc::scoped_refptr<const webrtc::RTCStatsReport>& report)>
+      ResultCallback;
+
+  static RTCStatsCallback* Create(ResultCallback result_callback) {
+    return new rtc::RefCountedObject<RTCStatsCallback>(
+        std::move(result_callback));
+  }
+
+  void OnStatsDelivered(
+      const rtc::scoped_refptr<const webrtc::RTCStatsReport>& report) override {
+    std::move(result_callback_)(report);
+  }
+
+ protected:
+  RTCStatsCallback(ResultCallback result_callback)
+      : result_callback_(std::move(result_callback)) {}
+  ~RTCStatsCallback() override = default;
+
+ private:
+  ResultCallback result_callback_;
+};
+
 RTCConnection::~RTCConnection() {
   connection_->Close();
 }
@@ -151,6 +177,12 @@ bool RTCConnection::isMediaEnabled(
     return track->enabled();
   }
   return false;
+}
+
+void RTCConnection::getStats(
+    std::function<void(const rtc::scoped_refptr<const webrtc::RTCStatsReport>&)>
+        callback) {
+  connection_->GetStats(RTCStatsCallback::Create(std::move(callback)));
 }
 
 }  // namespace sora