+++ src/sora.cpp
@@ -48,10 +48,11 @@ void Sora::DispatchEvents() {
   }
 }
 
-bool Sora::Connect(std::string signaling_url,
+bool Sora::Connect(std::string unity_version,
+                   std::string signaling_url,
                    std::string channel_id,
                    std::string metadata,
-                   bool downstream,
+                   std::string role,
                    bool multistream,
                    int capturer_type,
                    void* unity_camera_texture,
@@ -69,10 +70,10 @@ bool Sora::Connect(std::string signaling_url,
   signaling_url_ = std::move(signaling_url);
   channel_id_ = std::move(channel_id);
 
-  RTC_LOG(LS_INFO) << "Sora::Connect signaling_url=" << signaling_url_
+  RTC_LOG(LS_INFO) << "Sora::Connect unity_version=" << unity_version
+                   << " signaling_url =" << signaling_url_
                    << " channel_id=" << channel_id_ << " metadata=" << metadata
-                   << " downstream=" << downstream
-                   << " multistream=" << multistream
+                   << " role=" << role << " multistream=" << multistream
                    << " capturer_type=" << capturer_type
                    << " unity_camera_texture=0x" << unity_camera_texture
                    << " video_capturer_device=" << video_capturer_device
@@ -82,6 +83,12 @@ bool Sora::Connect(std::string signaling_url,
                    << " unity_audio_output=" << unity_audio_output
                    << " audio_recording_device=" << audio_recording_device
                    << " audio_playout_device=" << audio_playout_device;
+
+  if (role != "upstream" && role != "downstream" && role != "sendonly" &&
+      role != "recvonly" && role != "sendrecv") {
+    RTC_LOG(LS_ERROR) << "Invalid role: " << role;
+    return false;
+  }
 
   renderer_.reset(new UnityRenderer(
       [this](ptrid_t track_id) {
@@ -111,8 +118,12 @@ bool Sora::Connect(std::string signaling_url,
     return false;
   }
 
-  if (!downstream) {
-    // 送信側は capturer を設定する。playout の設定はしない
+  RTCManagerConfig config;
+  config.audio_recording_device = audio_recording_device;
+  config.audio_playout_device = audio_playout_device;
+
+  if (role == "upstream" || role == "sendonly" || role == "sendrecv") {
+    // 送信側は capturer を設定する。送信のみの場合は playout の設定はしない
     rtc::scoped_refptr<ScalableVideoTrackSource> capturer = CreateVideoCapturer(
         capturer_type, unity_camera_texture, video_capturer_device, video_width,
         video_height, unity_camera_capturer_);
@@ -120,10 +131,7 @@ bool Sora::Connect(std::string signaling_url,
       return false;
     }
 
-    RTCManagerConfig config;
-    config.no_playout = true;
-    config.audio_recording_device = audio_recording_device;
-    config.audio_playout_device = audio_playout_device;
+    config.no_playout = role == "upstream" || role == "sendonly";
     rtc_manager_ =
         RTCManager::Create(config, std::move(capturer), renderer_.get(),
                            unity_adm_, std::move(task_queue_factory));
@@ -132,8 +140,6 @@ bool Sora::Connect(std::string signaling_url,
     RTCManagerConfig config;
     config.no_recording = true;
     config.no_video = true;
-    config.audio_recording_device = audio_recording_device;
-    config.audio_playout_device = audio_playout_device;
     rtc_manager_ =
         RTCManager::Create(config, nullptr, renderer_.get(), unity_adm_,
                            std::move(task_queue_factory));
@@ -143,8 +149,16 @@ bool Sora::Connect(std::string signaling_url,
     RTC_LOG(LS_INFO) << "Start Signaling: url=" << signaling_url_
                      << " channel_id=" << channel_id_;
     SoraSignalingConfig config;
-    config.role = downstream ? SoraSignalingConfig::Role::Downstream
-                             : SoraSignalingConfig::Role::Upstream;
+    config.unity_version = unity_version;
+    config.role = role == "upstream"
+                      ? SoraSignalingConfig::Role::Upstream
+                      : role == "downstream"
+                            ? SoraSignalingConfig::Role::Downstream
+                            : role == "sendonly"
+                                  ? SoraSignalingConfig::Role::Sendonly
+                                  : role == "recvonly"
+                                        ? SoraSignalingConfig::Role::Recvonly
+                                        : SoraSignalingConfig::Role::Sendrecv;
     config.multistream = multistream;
     config.signaling_url = signaling_url_;
     config.channel_id = channel_id_;
@@ -210,10 +224,9 @@ int Sora::GetRenderCallbackEventID() const {
 }
 
 void Sora::RenderCallback() {
-  if (!unity_camera_capturer_) {
-    return;
+  if (unity_camera_capturer_) {
+    unity_camera_capturer_->OnRender();
   }
-  unity_camera_capturer_->OnRender();
 }
 void Sora::ProcessAudio(const void* p, int offset, int samples) {
   if (!unity_adm_) {