+++ src/sora.h
@@ -83,10 +83,11 @@ class Sora {
   void SetOnNotify(std::function<void(std::string)> on_notify);
   void DispatchEvents();
 
-  bool Connect(std::string signaling_url,
+  bool Connect(std::string unity_version,
+               std::string signaling_url,
                std::string channel_id,
                std::string metadata,
-               bool downstream,
+               std::string role,
                bool multistream,
                int capturer_type,
                void* unity_camera_texture,