+++ CHANGES.md
@@ -11,23 +11,90 @@
 
 ## develop
 
-**変更履歴は 1.0.0 から**
+## 2020.1
 
-- [ADD] Unity のカメラ映像を取得し Sora で送信
-- [ADD] カメラから映像を取得し Sora に送信
+- [ADD] Windows 版では H.264 エンコードに NVIDIA VIDEO CODEC SDK を利用する（利用可能な場合のみ）
+    - 実験的機能
+    - @melpon
+- [ADD] macOS 版では H.264 のエンコード/デコードに VideoToolbox を利用する
+    - 実験的機能
+    - @melpon
+- [ADD] Sora へ WebRTC 統計情報を送るようにした
+    - @melpon
+
+## 1.0.4
+
+- [UPDATE] webrtc-build を 80.3987.2.2 に上げた
+    - @melpon
+- [CHANGE] Sora への接続時に sendonly, recvonly, sendrecv を指定できるようにした
+    - @melpon
+
+## 1.0.3
+
+- [UPDATE] Boost のバージョンを 1.72.0 に上げた
+    - @melpon
+- [UPDATE] WebRTC のバージョンを M80 (3987@{#2}) に上げた
+    - @melpon
+
+## 1.0.2
+
+- [UPDATE] WebRTC のバージョンを M79@{#5} に上げた
+    - @melpon
+- [UPDATE] Windows 版の WebRTC ライブラリを shiguredo-webrtc-build からダウンロードするようにした
+    - @melpon
+- [UPDATE] macOS 版の WebRTC ライブラリを shiguredo-webrtc-build からダウンロードするようにした
+    - @melpon
+
+## 1.0.1
+
+- [ADD] connect 時の文字列に SDK や WebRTC のバージョンを送信する
+    - @melpon
+
+## 1.0.0
+
+- [ADD] Windows 10 x86_64 対応
+    - @melpon
+- [ADD] macOS 10.15 対応
+    - @melpon
+- [ADD] macOS x86_64 でのデバイス指定機能を追加
+    - @melpon
+- [ADD] Windows x86_64 でのデバイス指定機能を追加
+    - @melpon
+- [ADD] カメラから映像を取得し Sora で送信
+    - @melpon
 - [ADD] カメラから映像を取得し Unity アプリに出力
-- [ADD] マイクから音声を取得し Sora に送信
+    - @melpon
+- [ADD] マイクから音声を取得し Sora で送信
+    - @melpon
 - [ADD] マイクから音声を取得し Unity アプリに出力
+    - @melpon
 - [ADD] Unity アプリで Sora からの音声を受信
+    - @melpon
 - [ADD] Unity アプリで Sora からの映像を受信
-- [ADD] Unity アプリで Sora からの音声を再生
+    - @melpon
 - [ADD] Sora から受信した音声を Unity アプリに出力
+    - @melpon
 - [ADD] Sora から受信した映像を Unity アプリに出力
-- [ADD] ソフトウェアエンコード/デコード VP8 / VP9 への対応
+    - @melpon
+- [ADD] マルチストリームへの対応
+    - @melpon
+- [ADD] VP8 / VP9 への対応
+    - @melpon
 - [ADD] Opus への対応
-- [ADD] Windows 10 x86_64 への対応
-- [ADD] macOS x86_64 への対応
-- [ADD] Sora の音声ビットレート指定へ対応
-- [ADD] Sora の映像ビットレート指定へ対応
-- [ADD] Sora のマルチストリームへ対応
-
+    - @melpon
+- [ADD] Unity カメラからの映像取得に対応
+    - @melpon
+- [ADD] Sora のシグナリング通知に対応
+    - @melpon
+- [ADD] Sora の metadata に対応
+    - @melpon
+- [ADD] マイクの代わりに Unity からのオーディを出力に対応
+    - @melpon
+- [ADD] シグナリング開始時のコーデック指定機能を追加
+    - @melpon
+- [ADD] シグナリング開始時のビットレート指定機能を追加
+    - @melpon
+- [ADD] Unity 側で受信したオーディオを再生できるようにする
+    - @melpon
+- [ADD] 受信したオーディオを AudioClip から再生できるようにする
+    - @melpon